Refactor vehicle component structure and implement CarForm and CarList for car registration and display
